+++ docs/source/release_notes.rst
@@ -7,9 +7,10 @@ Release 0.3 (October 2019)
 ----------------------------
 * Added wrapper class for different backends
 * Added documentation
+* Added (non-optimized, single core) Python implementation 
 
 Release 0.2 (July 2019)
 ----------------------------
 * Initial release
-* Basic implementation for BFAST-Monitor
+* Basic implementation for BFASTMonitor
 * Multi-OpenCL-device support